NET-2: renamed hostnetwork to dhtnetwork, minor fixes in servicenetwork
