+++ network/dhtnetwork/dht.go
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package hostnetwork
+package dhtnetwork
 
 import (
 	"bytes"
@@ -31,10 +31,10 @@ import (
 	"github.com/insolar/insolar/log"
 	"github.com/insolar/insolar/metrics"
 	"github.com/insolar/insolar/network/consensus"
-	"github.com/insolar/insolar/network/hostnetwork/hosthandler"
-	"github.com/insolar/insolar/network/hostnetwork/resolver"
-	"github.com/insolar/insolar/network/hostnetwork/routing"
-	"github.com/insolar/insolar/network/hostnetwork/store"
+	"github.com/insolar/insolar/network/dhtnetwork/hosthandler"
+	"github.com/insolar/insolar/network/dhtnetwork/resolver"
+	"github.com/insolar/insolar/network/dhtnetwork/routing"
+	"github.com/insolar/insolar/network/dhtnetwork/store"
 	"github.com/insolar/insolar/network/transport"
 	"github.com/insolar/insolar/network/transport/host"
 	"github.com/insolar/insolar/network/transport/id"
