+++ network/dhtnetwork/store/memorystore_test.go
